+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [1.1.0] -2021-09-30
 ### Added
 - Add `tsrpc.config.ts`, update `npm run dev` and `npm run build`
-- Update `tsrpc-cli` verstion to 2.0.6
+- Update `tsrpc-cli` verstion to `^2.0.8`
 ### Changed
 - 改用 `chalk` 替代 `colors`