+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## [1.1.0] -2021-09-30
 ### Added
 - Add `tsrpc.config.ts`, update `npm run dev` and `npm run build`
-- Update `tsrpc-cli` verstion to `^2.0.8`
+- Update `tsrpc-cli` version to `^2.0.8`
+- 为 Vue 项目增加 `vetur.config.js`
 ### Changed
 - 改用 `chalk` 替代 `colors`